+++ forwarder/forwarder.js
@@ -80,28 +80,31 @@ module.exports = function (RED)
         let handleTopic = msg =>
         {
             let real_topic = helper.getTopicName(msg.topic);
-
-            if (real_topic.startsWith("set_state"))
-                real_topic = real_topic.replace("set_state", "set");
-
-            if (real_topic == "set_inverted")
-            {
-                real_topic = "set";
-                msg.payload = !msg.payload;
-            }
-
             let new_state = null;
-            if (real_topic == "enable" || (real_topic == "set" && msg.payload))
-                new_state = true;
-            else if (real_topic == "disable" || (real_topic == "set" && !msg.payload))
-                new_state = false;
+
+            if (real_topic != null)
+            {
+                if (real_topic.startsWith("set_state"))
+                    real_topic = real_topic.replace("set_state", "set");
+
+                if (real_topic == "set_inverted")
+                {
+                    real_topic = "set";
+                    msg.payload = !msg.payload;
+                }
+
+                if (real_topic == "enable" || (real_topic == "set" && msg.payload))
+                    new_state = true;
+                else if (real_topic == "disable" || (real_topic == "set" && !msg.payload))
+                    new_state = false;
+            }
 
             switch (real_topic)
             {
                 case "debug":
                     helper.nodeDebug(node, {
                         node_settings,
-                        forward_true,                        
+                        forward_true,
                         forward_false,
                         forward_last_on_enable,
                     });